HW1/requirements.txt
torch
torchvision
pandas
Pillow
tqdm
numpy==1.23
seaborn
matplotlib